Second reorganization to solve cyclic-imports problems
