+++ xclim/__init__.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 """Climate indices computation package based on xArray"""
+from xclim.core import units
 from xclim.indicators import atmos
 from xclim.indicators import ICCLIM
 from xclim.indicators import land
 from xclim.indicators import seaIce
-
 
 __author__ = """Travis Logan"""
 __email__ = "logan.travis@ouranos.ca"
