+++ xclim/core/calendar.py
