+++ xclim/core/checks.py
