+++ xclim/core/utils.py
@@ -6,10 +6,46 @@ Miscellaneous indices utilities
 Helper functions for the indices computation, things that do not belong in neither
 `xclim.indices.calendar`, `xclim.indices.fwi`, `xclim.indices.generic` or `xclim.indices.run_length`.
 """
+from collections import defaultdict
+from types import FunctionType
+
 import numpy as np
 import xarray as xr
 
-from xclim.units import convert_units_to
+from xclim.core.units import convert_units_to
+
+
+def wrapped_partial(func: FunctionType, *args, **kwargs):
+    from functools import partial, update_wrapper
+
+    partial_func = partial(func, *args, **kwargs)
+    update_wrapper(partial_func, func)
+    return partial_func
+
+
+# TODO Reconsider the utility of this
+def walk_map(d: dict, func: FunctionType):
+    """Apply a function recursively to values of dictionary.
+
+    Parameters
+    ----------
+    d : dict
+      Input dictionary, possibly nested.
+    func : FunctionType
+      Function to apply to dictionary values.
+
+    Returns
+    -------
+    dict
+      Dictionary whose values are the output of the given function.
+    """
+    out = {}
+    for k, v in d.items():
+        if isinstance(v, (dict, defaultdict)):
+            out[k] = walk_map(v, func)
+        else:
+            out[k] = func(v)
+    return out
 
 
 def uas_vas_2_sfcwind(uas: xr.DataArray = None, vas: xr.DataArray = None):
