+++ xclim/indicators/land/_streamflow.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 import numpy as np
 
-from xclim.indicators import checks
-from xclim.indicators import Indicator
-from xclim.indicators.utils import wrapped_partial
+from xclim.core import checks
+from xclim.core.indicator import Indicator
+from xclim.core.utils import wrapped_partial
 from xclim.indices import base_flow_index
 from xclim.indices import generic
