+++ xclim/indicators/seaIce/_seaice.py
@@ -5,8 +5,8 @@ Sea ice indicators
 import abc
 
 from xclim import indices
-from xclim.indicators import checks
-from xclim.indicators import Indicator2D
+from xclim.core import checks
+from xclim.core.indicator import Indicator2D
 
 __all__ = ["sea_ice_area", "sea_ice_extent"]
