--- xclim/units/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-`Pint` is used to define the `units` `UnitRegistry` and `xclim.units.core` defines
-most unit handling methods.
-"""
-from ._registry import units
-from .core import *
